@@ -1,11 +1,11 @@
 name: views_weebi_example
 description: A new Flutter package project.
 version: 0.0.1
-homepage:
+homepage: weebi.com
 publish_to: none
 
 environment:
-  sdk: '>=2.7.0 <3.0.0' #  sdk: '>=2.18.4 <3.0.0'
+  sdk: '>=2.19.6 <3.0.0' #  sdk: '>=2.18.4 <3.0.0'
   flutter: '>=1.17.0'
 
 dependencies:
@@ -19,10 +19,10 @@ dependencies:
 
   views_weebi:
     path: ../../views
-  models_weebi: ^1.0.13
-    #path: ../models_weebi 
-  mixins_weebi: ^0.0.9
-    #path: ../mixins_weebi
+  models_weebi: #^1.0.13
+    path: ../../models 
+  mixins_weebi: #^0.0.9
+    path: ../../mixins
  
 
 dev_dependencies:
