@@ -5,7 +5,7 @@ description: views and widgets
 #publish_to: none #https://onepub.dev/api/xnxpsdavuh/
 
 environment: 
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.19.6 <3.0.0' #'>=2.7.0 <3.0.0'
   flutter: '>=1.17.0'
 dependencies: 
   flutter: 
@@ -17,12 +17,12 @@ dependencies:
   provider: ^6.0.3
   rc_router2: ^1.0.0
   flutter_typeahead: 3.1.0 #4.0.0
-  flutter_styled_toast: 1.5.2+3 #^2.1.3
+  flutter_styled_toast: ^2.1.3 #1.5.2+3 
 
-  models_weebi: ^1.0.13
-    #path: ../models_weebi 
-  mixins_weebi: ^0.0.9
-    #path: ../mixins_weebi
+  models_weebi: #^1.0.13
+    path: ../models
+  mixins_weebi: #^0.0.9
+    path: ../mixins
 
 dev_dependencies:
   flutter_test: 
