@@ -5,14 +5,14 @@ import 'import/import_articles_json.dart';
 
 class InfoView extends StatelessWidget {
   final GlobalKey<NavigatorState> mainNavigator;
-  const InfoView(this.mainNavigator, {Key key}) : super(key: key);
+  const InfoView(this.mainNavigator, {Key? key}) : super(key: key);
   // static const routePath = '/info_view';
   @override
   Widget build(BuildContext context) {
     return ChassisTutoProducts(
       selectedIndex: 1,
       mainNavigatorKey: mainNavigator,
-      body: const ImportArticleLineJsonView(),
+      body: const ImportArticleCalibreJsonView(),
     );
   }
 }
