@@ -2,51 +2,19 @@
 import 'package:flutter/material.dart';
 
 // Package imports:
-import 'package:mixins_weebi/mobx_store_closing.dart';
 import 'package:models_weebi/utils.dart';
-import 'package:provider/provider.dart';
 
 // Project imports:
-import 'package:models_weebi/weebi_models.dart' show ArticleBasket, ArticleLine;
 import 'package:mixins_weebi/stock.dart';
 
 import 'package:views_weebi/routes.dart';
-import 'package:mixins_weebi/stores.dart' show ArticlesStore;
-import 'package:mixins_weebi/stores.dart' show TicketsStore;
 import 'package:views_weebi/src/articles/photo.dart';
 import 'package:views_weebi/styles.dart' show WeebiColors;
 import 'package:views_weebi/views_article_basket.dart';
 
-class ArticleBasketGlimpseWidget extends ArticleStockStatelessAbstract
-    with ArticleBasketRealizableNow {
-  final ArticleLine line;
-
-  const ArticleBasketGlimpseWidget(this.line, ArticleBasket article)
-      : super(article);
-
-  @override
-  Widget build(BuildContext context) {
-    final ticketsStore = Provider.of<TicketsStore>(context, listen: false);
-    final closingsStore = Provider.of<ClosingsStore>(context, listen: false);
-    final articlesStore = Provider.of<ArticlesStore>(context, listen: false);
-
-    final doTheGroove = articleBasketWrapThem(
-        closingsStore.closingStockShops ?? [],
-        ticketsStore.tickets,
-        articlesStore.lines,
-        WeebiDates.defaultFirstDate,
-        DateTime.now());
-    int realizableBaskets = basketsRealizableNow(doTheGroove);
-
-    return ArticleBasketGlimpseWidStateFul(article, realizableBaskets);
-  }
-}
-
 class ArticleBasketGlimpseWidStateFul extends StatefulWidget {
-  final ArticleBasket article;
-  final int realizableBaskets;
-  const ArticleBasketGlimpseWidStateFul(this.article, this.realizableBaskets,
-      {Key key})
+  final StockNowArticleBasket articleStock;
+  const ArticleBasketGlimpseWidStateFul(this.articleStock, {Key? key})
       : super(key: key);
 
   @override
@@ -56,19 +24,15 @@ class ArticleBasketGlimpseWidStateFul extends StatefulWidget {
 
 class _ArticleBasketGlimpseWidStateFulState
     extends State<ArticleBasketGlimpseWidStateFul> {
-  Color iconColor;
-  @override
-  void initState() {
-    super.initState();
-    iconColor = Colors.grey;
-  }
+  Color iconColor = Colors.grey;
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
       onLongPress: () {
         Navigator.of(context).pushNamed(ArticleDetailRoute.generateRoute(
-            '${widget.article.lineId}', '${widget.article.id}'));
+            '${widget.articleStock.article.calibreId}',
+            '${widget.articleStock.article.id}'));
       },
       child: ExpansionTile(
         onExpansionChanged: (bool expanding) => setState(() =>
@@ -82,34 +46,36 @@ class _ArticleBasketGlimpseWidStateFulState
         ),
         title: Row(
           children: <Widget>[
-            widget.article.photo != null && widget.article.photo.isNotEmpty
+            widget.articleStock.article.photo.isNotEmpty
                 ? Hero(
-                    tag: '${widget.article.lineId}.${widget.article.id}',
+                    tag:
+                        '${widget.articleStock.article.calibreId}.${widget.articleStock.article.id}',
                     child: CircleAvatar(
-                        foregroundImage: PhotoWidget(widget.article).getImage
-                            as ImageProvider),
+                        foregroundImage:
+                            PhotoWidget(widget.articleStock.article).getImage
+                                as ImageProvider),
                   )
                 : CircleAvatar(backgroundColor: Colors.transparent),
             Padding(
               padding: const EdgeInsets.symmetric(horizontal: 4.0),
               child: Text(
-                '#${widget.article.lineId}.${widget.article.id}',
+                '#${widget.articleStock.article.calibreId}.${widget.articleStock.article.id}',
                 style:
                     const TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
               ),
             ),
             Expanded(
               child: Text(
-                '${widget.article.fullName}',
-                style: (widget.article.status ?? true) == false
+                '${widget.articleStock.article.fullName}',
+                style: (widget.articleStock.article.status) == false
                     ? const TextStyle(decoration: TextDecoration.lineThrough)
                     : const TextStyle(),
               ),
             ),
             Padding(
               padding: const EdgeInsets.all(8.0),
-              child: Text(widget.realizableBaskets > 0
-                  ? '${numFormat?.format(widget.realizableBaskets)}' // réalisables
+              child: Text(widget.articleStock.stockRealizablesBasketsNow > 0
+                  ? '${numFormat.format(widget.articleStock.stockRealizablesBasketsNow)}' // réalisables
                   : ''),
             ),
             // SizedBox(width: ScreenUtil().setWidth(10)),
@@ -117,7 +83,7 @@ class _ArticleBasketGlimpseWidStateFulState
           ],
         ),
         children: <Widget>[
-          for (final proxy in (widget.article).proxies)
+          for (final proxy in (widget.articleStock.article).proxies)
             ProxyArticleGlimpseFrameWidget(proxy)
         ],
       ),
