@@ -8,22 +8,28 @@ import 'package:views_weebi/src/articles/article/article_basket/proxy_glimpse.da
 
 class ProxyArticleGlimpseFrameWidget extends StatelessWidget {
   final ProxyArticle proxy;
-  const ProxyArticleGlimpseFrameWidget(this.proxy, {Key key}) : super(key: key);
+  const ProxyArticleGlimpseFrameWidget(this.proxy, {Key? key})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     final articlesStore = Provider.of<ArticlesStore>(context, listen: false);
-    final _line =
-        articlesStore.lines.firstWhereOrNull((l) => l.id == proxy.proxyLineId);
-    if (_line == null) {
-      print('_line not found for proxy.proxyLineId ${proxy.proxyLineId}');
+    final _calibre = articlesStore.calibres
+        .firstWhereOrNull((l) => l.id == proxy.proxyCalibreId);
+    if (_calibre == null) {
+      print('_calibre not found for proxy.proxyLineId ${proxy.proxyCalibreId}');
     }
-    final _article = _line.articles.firstWhere(
-        (a) => a.lineId == proxy.proxyLineId && a.id == proxy.proxyArticleId);
+    if (_calibre?.articles == null) {
+      print(
+          '_calibre?.articles empty for proxy.proxyLineId ${proxy.proxyCalibreId}');
+    }
+    final _article = _calibre?.articles.firstWhere((a) =>
+        a.calibreId == proxy.proxyCalibreId && a.id == proxy.proxyArticleId);
     if (_article == null) {
       print(
-          '_article not found for proxy.proxyLineId ${proxy.proxyLineId} && proxy.proxyArticleId ${proxy.proxyArticleId}');
+          '_article not found for proxy.proxyLineId ${proxy.proxyCalibreId} && proxy.proxyArticleId ${proxy.proxyArticleId}');
     }
-    return ProxyAGlimpseWidget(article: _article, proxy: proxy);
+    return ProxyAGlimpseWidget(
+        article: _article as ArticleRetail, proxy: proxy);
   }
 }
