@@ -3,36 +3,35 @@ import 'package:flutter/material.dart';
 import 'package:models_weebi/utils.dart';
 
 // Project imports:
-import 'package:models_weebi/weebi_models.dart' show ArticleLine;
+import 'package:models_weebi/weebi_models.dart' show ArticleBasket;
 import 'package:views_weebi/src/articles/photo.dart';
 
 class LineSingleArticleBasketTileTitle extends StatelessWidget {
-  final ArticleLine line;
+  final ArticleBasket articleBasket;
   final int realizableBaskets;
   final Color iconColor;
   const LineSingleArticleBasketTileTitle(
-      this.line, this.realizableBaskets, this.iconColor,
-      {Key key})
+      this.articleBasket, this.realizableBaskets, this.iconColor,
+      {Key? key})
       : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Row(
       children: <Widget>[
-        line?.articles?.first?.photo == null ||
-                line.articles.first.photo.isEmpty
+        articleBasket.photo.isEmpty
             ? CircleAvatar(backgroundColor: Colors.transparent)
             : Hero(
-                tag: line.id,
+                tag: articleBasket.calibreId,
                 child: CircleAvatar(
-                  foregroundImage: PhotoWidget(line.articles.first).getImage
-                      as ImageProvider,
+                  foregroundImage:
+                      PhotoWidget(articleBasket).getImage as ImageProvider,
                 ),
               ),
         Padding(
           padding: const EdgeInsets.symmetric(horizontal: 4.0),
           child: Text(
-            '#${line.id}',
+            '#${articleBasket.calibreId}',
             style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 18),
           ),
         ),
@@ -40,8 +39,8 @@ class LineSingleArticleBasketTileTitle extends StatelessWidget {
           child: Padding(
             padding: const EdgeInsets.symmetric(horizontal: 8),
             child: Text(
-              '${line.title}', // (${line.stockUnitText})
-              style: line.articles.first.status == false
+              '${articleBasket.fullName}', // (${line.stockUnitText})
+              style: articleBasket.status == false
                   ? const TextStyle(decoration: TextDecoration.lineThrough)
                   : const TextStyle(),
             ),
@@ -52,7 +51,7 @@ class LineSingleArticleBasketTileTitle extends StatelessWidget {
             Padding(
               padding: const EdgeInsets.all(8.0),
               child: Text(realizableBaskets > 0
-                  ? '${numFormat?.format(realizableBaskets)}'
+                  ? '${numFormat.format(realizableBaskets)}'
                   : ''),
             ),
             Icon(Icons.shopping_basket, color: iconColor)
