@@ -2,16 +2,17 @@
 import 'package:flutter/material.dart';
 
 // Project imports:
-import 'package:models_weebi/weebi_models.dart' show ArticleLine;
+import 'package:models_weebi/weebi_models.dart' show ArticleCalibre;
 import 'package:models_weebi/utils.dart';
 import 'package:views_weebi/src/articles/photo.dart';
 
-class LineArticleRetailTileTitle extends StatelessWidget {
-  final ArticleLine line;
-  final double lineLiveQt;
+class ArticleRetailCalibreTileTitle extends StatelessWidget {
+  final ArticleCalibre articleCalibre;
+  final double calibreStockNow;
   final Color iconColor;
-  const LineArticleRetailTileTitle(this.line, this.lineLiveQt, this.iconColor,
-      {Key key})
+  const ArticleRetailCalibreTileTitle(
+      this.articleCalibre, this.calibreStockNow, this.iconColor,
+      {Key? key})
       : super(key: key);
 
   @override
@@ -22,14 +23,14 @@ class LineArticleRetailTileTitle extends StatelessWidget {
           height: 42,
           width: 42,
           child: Hero(
-            tag: '${line.id}.${line.articles.first.id}',
-            child: ClipOval(child: PhotoWidget(line.articles.first)),
+            tag: '${articleCalibre.id}.${articleCalibre.articles.first.id}',
+            child: ClipOval(child: PhotoWidget(articleCalibre.articles.first)),
           ),
         ),
         Padding(
           padding: const EdgeInsets.symmetric(horizontal: 4.0),
           child: Text(
-            '#${line.id}',
+            '#${articleCalibre.id}',
             style: const TextStyle(
               fontWeight: FontWeight.bold,
               fontSize: 18,
@@ -40,19 +41,19 @@ class LineArticleRetailTileTitle extends StatelessWidget {
           child: Padding(
             padding: const EdgeInsets.symmetric(horizontal: 8),
             child: Text(
-              line.title, // (${line.stockUnitText})
-              style: line.articles.first.status == false
+              articleCalibre.title,
+              style: articleCalibre.articles.first.status == false
                   ? const TextStyle(decoration: TextDecoration.lineThrough)
                   : const TextStyle(),
             ),
           ),
         ),
-        if (line.title != '*') //
+        if (articleCalibre.title != '*') //
           Row(
             children: [
               Padding(
                 padding: const EdgeInsets.all(8.0),
-                child: Text(numFormat.format(lineLiveQt)),
+                child: Text(numFormat.format(calibreStockNow)),
               ),
               Icon(Icons.warehouse, color: iconColor),
             ],
