@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 abstract class ChassisAbstract extends StatefulWidget {
-  static const Key keyAppBarTitle = Key("AppBar");
+  static const Key? keyAppBarTitle = Key("AppBar");
   final Widget body;
   final int selectedIndex;
   final GlobalKey<NavigatorState> mainNavigatorKey;
@@ -9,9 +9,9 @@ abstract class ChassisAbstract extends StatefulWidget {
   final List<Widget> actions;
   const ChassisAbstract(
       {key,
-      @required this.body,
-      @required this.selectedIndex,
-      @required this.mainNavigatorKey,
-      @required this.floatingButton,
-      @required this.actions});
+      required this.body,
+      required this.selectedIndex,
+      required this.mainNavigatorKey,
+      required this.floatingButton,
+      required this.actions});
 }
