@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class IconAddArticleBasket extends StatelessWidget {
-  const IconAddArticleBasket({Key key}) : super(key: key);
+  const IconAddArticleBasket({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
