@@ -9,40 +9,38 @@ import 'package:provider/provider.dart';
 import 'package:views_weebi/buttons.dart';
 
 class WeebiOkPaddedQtDouble extends StatelessWidget {
-  const WeebiOkPaddedQtDouble({Key key}) : super(key: key);
+  const WeebiOkPaddedQtDouble({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     final _qtEditVal =
         Provider.of<AmountProvider>(context, listen: true).textEditVal;
-    return WeebiButtonOkRectWide(
-      tooltip: 'valider',
-      string: 'OK',
-      onPressed: (double.tryParse(_qtEditVal.text) ?? 0) < 0.01
-          ? null
-          : () {
-              Navigator.of(context)
-                  .pop((double.tryParse(_qtEditVal.text) ?? 0.0));
-            },
-    );
+    return ((double.tryParse(_qtEditVal.text) ?? 0) < 0.01)
+        ? const SizedBox()
+        : WeebiButtonOkRectWide(
+            tooltip: 'valider',
+            string: 'OK',
+            onPressed: () => Navigator.of(context)
+                .pop((double.tryParse(_qtEditVal.text) ?? 0.0)),
+          );
   }
 }
 
 class WeebiOkPaddedInt extends StatelessWidget {
-  const WeebiOkPaddedInt({Key key}) : super(key: key);
+  const WeebiOkPaddedInt({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     final _qtEditVal =
         Provider.of<AmountProvider>(context, listen: true).textEditVal;
-    return WeebiButtonOkRectWide(
-      tooltip: 'valider',
-      string: 'OK',
-      onPressed: (int.tryParse(_qtEditVal.text) ?? 0) < 1
-          ? null
-          : () {
+    return ((double.tryParse(_qtEditVal.text) ?? 0) < 1)
+        ? const SizedBox()
+        : WeebiButtonOkRectWide(
+            tooltip: 'valider',
+            string: 'OK',
+            onPressed: () {
               Navigator.of(context).pop((int.tryParse(_qtEditVal.text) ?? 0));
             },
-    );
+          );
   }
 }
