@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:views_weebi/views_line.dart';
-import '../lib/src/1_providers.dart';
-import '../lib/src/2_stores_loader.dart';
-import '../lib/src/3_material_app.dart';
-import '../lib/src/line_dummy.dart';
+import 'package:views_weebi/views_calibre.dart';
+import 'package:views_weebi_example/src/1_providers.dart';
+import 'package:views_weebi_example/src/2_stores_loader.dart';
+import 'package:views_weebi_example/src/3_material_app.dart';
+import 'package:views_weebi_example/src/article_calibre_dummy.dart';
 
 void main() {
   testWidgets('check articles search by title', (tester) async {
-    // pumps ChassisTutoProducts and also ArticlesLinesViewWIP
+    // pumps ChassisTutoProducts and also ArticlesCalibresViewWIP
     await tester.pumpWidget(
-        ProvidersW(StoresLoader(const ExampleApp(), articleLinesDummies)));
+        ProvidersW(StoresLoader(const ExampleApp(), articleCalibresDummies)));
     await tester.pump();
     expect(find.text('Articles'), findsWidgets);
 
@@ -24,15 +24,15 @@ void main() {
     await tester.enterText(searchBar, 'ba');
     await tester.pump();
     int itemsLength = tester
-        .widgetList<ArticleLineFrame>(find.byType(ArticleLineFrame))
+        .widgetList<ArticleCalibreFrame>(find.byType(ArticleCalibreFrame))
         .length;
     expect(itemsLength, 1);
 
-    ArticleLineFrame lineFrameViewBabibel = tester
-        .widgetList<ArticleLineFrame>(find.byType(ArticleLineFrame))
+    ArticleCalibreFrame articleCalibreFrameViewBabibel = tester
+        .widgetList<ArticleCalibreFrame>(find.byType(ArticleCalibreFrame))
         .elementAt(0);
-    expect(lineFrameViewBabibel.line.title, 'Babibel');
-    expect(lineFrameViewBabibel.line.id, 2);
+    expect(articleCalibreFrameViewBabibel.calibre.title, 'Babibel');
+    expect(articleCalibreFrameViewBabibel.calibre.id, 2);
 
     // verify that close icon is there and works
     final closeIcon = find.byTooltip('Chercher un article');
@@ -41,13 +41,13 @@ void main() {
     await tester.pump();
     expect(searchBar, findsNothing);
 
-    final linesSearchClose =
-        tester.widgetList<ArticleLineFrame>(find.byType(ArticleLineFrame));
+    final linesSearchClose = tester
+        .widgetList<ArticleCalibreFrame>(find.byType(ArticleCalibreFrame));
     expect(linesSearchClose.isNotEmpty, isTrue);
 
-    final ArticleLineFrame lineFrameView2 = linesSearchClose.elementAt(0);
-    expect(lineFrameView2.line.articles.length, 3);
-    expect(lineFrameView2.line.title, 'Noix de cola');
-    expect(lineFrameView2.line.id, 1);
+    final ArticleCalibreFrame lineFrameView2 = linesSearchClose.elementAt(0);
+    expect(lineFrameView2.calibre.articles.length, 3);
+    expect(lineFrameView2.calibre.title, 'Noix de cola');
+    expect(lineFrameView2.calibre.id, 1);
   });
 }
