@@ -1,17 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:views_weebi/views_line.dart';
-import '../lib/src/1_providers.dart';
-import '../lib/src/2_stores_loader.dart';
-import '../lib/src/3_material_app.dart';
-import '../lib/src/line_dummy.dart';
+import 'package:views_weebi/views_calibre.dart';
+import 'package:views_weebi_example/src/1_providers.dart';
+import 'package:views_weebi_example/src/2_stores_loader.dart';
+import 'package:views_weebi_example/src/3_material_app.dart';
+import 'package:views_weebi_example/src/article_calibre_dummy.dart';
 
 void main() {
   testWidgets('check articles sorting by title and codeShortcut',
       (tester) async {
     await tester.pumpWidget(
-        ProvidersW(StoresLoader(const ExampleApp(), articleLinesDummies)));
-    // pumps ChassisTutoProducts and also ArticlesLinesViewWIP
+        ProvidersW(StoresLoader(const ExampleApp(), articleCalibresDummies)));
+    // pumps ChassisTutoProducts and also ArticlesCalibresViewWIP
     await tester.pump();
 
     expect(find.text('Articles'), findsWidgets);
@@ -22,43 +22,43 @@ void main() {
     expect(find.byIcon(Icons.keyboard_arrow_up), findsNothing);
 
     // by default Noix de Cola is first
-    ArticleLineFrame item = tester
-        .widgetList<ArticleLineFrame>(find.byType(ArticleLineFrame))
+    ArticleCalibreFrame item = tester
+        .widgetList<ArticleCalibreFrame>(find.byType(ArticleCalibreFrame))
         .elementAt(0);
-    expect(item.line.title, 'Noix de cola');
-    expect(item.line.id, 1);
+    expect(item.calibre.title, 'Noix de cola');
+    expect(item.calibre.id, 1);
 
     // Tap the icon and trigger a frame
     // and verify that icon is updated and list reordered
     await tester.tap(sortByCodeFound);
     await tester.pump();
     expect(find.byIcon(Icons.keyboard_arrow_up), findsOneWidget);
-    ArticleLineFrame itemCodeAfterReorder = tester
-        .widgetList<ArticleLineFrame>(find.byType(ArticleLineFrame))
+    ArticleCalibreFrame itemCodeAfterReorder = tester
+        .widgetList<ArticleCalibreFrame>(find.byType(ArticleCalibreFrame))
         .elementAt(0);
-    expect('Sucre', itemCodeAfterReorder.line.title);
-    expect(3, itemCodeAfterReorder.line.id);
+    expect('Sucre', itemCodeAfterReorder.calibre.title);
+    expect(3, itemCodeAfterReorder.calibre.id);
 
     // and verify that icon is updated and list reordered back
     await tester.tap(sortByCodeFound);
     await tester.pump();
     expect(find.byIcon(Icons.keyboard_arrow_down), findsOneWidget);
     expect(find.byIcon(Icons.keyboard_arrow_up), findsNothing);
-    ArticleLineFrame itemCodeAfterRereorder = tester
-        .widgetList<ArticleLineFrame>(find.byType(ArticleLineFrame))
+    ArticleCalibreFrame itemCodeAfterRereorder = tester
+        .widgetList<ArticleCalibreFrame>(find.byType(ArticleCalibreFrame))
         .elementAt(0);
-    expect('Noix de cola', itemCodeAfterRereorder.line.title);
-    expect(1, itemCodeAfterRereorder.line.id);
+    expect('Noix de cola', itemCodeAfterRereorder.calibre.title);
+    expect(1, itemCodeAfterRereorder.calibre.id);
 
     // check sort by title works
     final sortByTitleFound = find.byTooltip('Trier par nom');
     expect(sortByTitleFound, findsOneWidget);
     await tester.tap(sortByTitleFound);
     await tester.pump();
-    ArticleLineFrame itemTitleAfterReorder = tester
-        .widgetList<ArticleLineFrame>(find.byType(ArticleLineFrame))
+    ArticleCalibreFrame itemTitleAfterReorder = tester
+        .widgetList<ArticleCalibreFrame>(find.byType(ArticleCalibreFrame))
         .elementAt(0);
-    expect('Babibel', itemTitleAfterReorder.line.title);
-    expect(2, itemTitleAfterReorder.line.id);
+    expect('Babibel', itemTitleAfterReorder.calibre.title);
+    expect(2, itemTitleAfterReorder.calibre.id);
   });
 }
