@@ -1,30 +1,30 @@
 //credits to Flutter Animation Gallery
 
 import 'package:flutter/material.dart';
-import 'package:mixins_weebi/stock.dart';
+import 'package:mixins_weebi/invokers.dart';
 import 'package:models_weebi/base.dart';
 import 'package:models_weebi/weebi_models.dart';
 
 import 'package:views_weebi/extensions.dart';
 import 'dart:math';
 
-import 'package:views_weebi/src/articles/article_card_slide.dart';
+import 'package:views_weebi/src/articles/article/article_retail/article_card_slide.dart';
 
 class SlidableCardsV2<A extends ArticleAbstract> extends StatefulWidget {
-  final ArticleLine line;
+  final ArticleCalibre calibre;
   final int articleId;
   final TicketsInvoker ticketsInvoker;
   final ClosingStockShopsInvoker closingStockShopsInvoker;
 
   const SlidableCardsV2(
-      this.line, this.ticketsInvoker, this.closingStockShopsInvoker,
+      this.calibre, this.ticketsInvoker, this.closingStockShopsInvoker,
       {this.articleId = 1});
   @override
   SlidableCardsV2State createState() => SlidableCardsV2State();
 }
 
 class SlidableCardsV2State extends State<SlidableCardsV2> {
-  PageController pageController;
+  PageController pageController = PageController();
   var currentPageValue = 0.0;
 
   @override
@@ -35,7 +35,7 @@ class SlidableCardsV2State extends State<SlidableCardsV2> {
     currentPageValue = (widget.articleId - 1).toDouble();
     pageController.addListener(() {
       setState(() {
-        currentPageValue = pageController.page;
+        currentPageValue = pageController.page!;
       });
     });
   }
@@ -56,7 +56,7 @@ class SlidableCardsV2State extends State<SlidableCardsV2> {
             physics: const BouncingScrollPhysics(
                 parent: AlwaysScrollableScrollPhysics()),
             controller: pageController,
-            itemCount: widget.line.articles.length,
+            itemCount: widget.calibre.articles.length,
             itemBuilder: (context, position) {
               if (position == currentPageValue) {
                 return Transform.scale(
@@ -64,8 +64,8 @@ class SlidableCardsV2State extends State<SlidableCardsV2> {
                   child: SizedBox(
                     width: 200,
                     child: ArticleACardSlide(
-                        widget.line,
-                        widget.line.articles[position],
+                        widget.calibre,
+                        widget.calibre.articles[position],
                         widget.ticketsInvoker,
                         widget.closingStockShopsInvoker),
                   ),
@@ -74,8 +74,8 @@ class SlidableCardsV2State extends State<SlidableCardsV2> {
                 return Transform.scale(
                   scale: max(1 - (currentPageValue - position), 0.5),
                   child: ArticleACardSlide(
-                      widget.line,
-                      widget.line.articles[position],
+                      widget.calibre,
+                      widget.calibre.articles[position],
                       widget.ticketsInvoker,
                       widget.closingStockShopsInvoker),
                 );
@@ -83,8 +83,8 @@ class SlidableCardsV2State extends State<SlidableCardsV2> {
                 return Transform.scale(
                   scale: max(1 - (position - currentPageValue), 0.5),
                   child: ArticleACardSlide(
-                      widget.line,
-                      widget.line.articles[position],
+                      widget.calibre,
+                      widget.calibre.articles[position],
                       widget.ticketsInvoker,
                       widget.closingStockShopsInvoker),
                 );
