@@ -1,35 +1,25 @@
 // Flutter imports:
 import 'package:flutter/material.dart';
+import 'package:mixins_weebi/stock.dart';
 
 // Package imports:
 
 // Project imports:
-import 'package:models_weebi/weebi_models.dart' show ArticleRetail;
-import 'package:mixins_weebi/stock.dart';
 
 import 'detail_section_article_retail.dart';
 import 'glimpse_a_retail.dart';
 
-class ArticleRetailFrameView
-    extends ArticleStockStatelessAbstract<ArticleRetail>
-    with ArticleStockNowMixin<ArticleRetail> {
+class ArticleRetailFrameView extends StatelessWidget {
   final bool isGlimpse;
-  final TicketsInvoker ticketsInvoker;
-  final ClosingStockShopsInvoker closingStockShopsInvoker;
-  const ArticleRetailFrameView(
-    ArticleRetail article,
-    this.isGlimpse,
-    this.ticketsInvoker,
-    this.closingStockShopsInvoker,
-  ) : super(article);
+  final StockNowArticleRetail articleStock;
+  const ArticleRetailFrameView(this.articleStock, this.isGlimpse);
 
   @override
   Widget build(BuildContext context) {
-    final double articleLiveQt =
-        articleStockNow(closingStockShopsInvoker.call(), ticketsInvoker.call());
-
     return isGlimpse
-        ? ArticleRetailGlimpseWidget(article, articleLiveQt)
-        : ArticleRetailDetailSection(article, articleLiveQt);
+        ? ArticleRetailGlimpseWidget(
+            articleStock.article, articleStock.stockNow)
+        : ArticleRetailDetailSection(
+            articleStock.article, articleStock.stockNow);
   }
 }
