@@ -11,28 +11,29 @@ import 'package:views_weebi/views_article.dart';
 import 'package:mixins_weebi/stores.dart';
 
 class ArticleDetailRoute extends RcRoute {
-  static String routePath = '/articles/lines/:lineId/:articleId';
+  static String routePath = '/articles/lines/:calibreId/:articleId';
 
-  static String generateRoute(String lineId, String articleId) =>
+  static String generateRoute(String calibreId, String articleId) =>
       RcRoute.generateRoute(routePath,
-          pathParams: {'lineId': lineId, 'articleId': articleId});
+          pathParams: {'calibreId': calibreId, 'articleId': articleId});
 
   ArticleDetailRoute() : super(path: ArticleDetailRoute.routePath);
 
   @override
   Widget build(BuildContext context) {
     final routeParams = Provider.of<RcRouteParameters>(context);
-    final lineId = routeParams.pathParameters['lineId'] ?? '';
+    final calibreId = routeParams.pathParameters['calibreId'] ?? '';
     final articleId = routeParams.pathParameters['articleId'] ?? '';
     final articlesStore = Provider.of<ArticlesStore>(context, listen: false);
-    final line = articlesStore.lines
-        .firstWhere((l) => l.id == int.tryParse(lineId), orElse: () {
-      throw 'no line match $lineId';
+    final calibre = articlesStore.calibres
+        .firstWhere((l) => l.id == int.tryParse(calibreId), orElse: () {
+      throw 'no calibre match $calibreId';
     });
     // this is not a article create route case
-    final article = line.articles.firstWhere(
-        (a) => '${a.lineId}' == lineId && '${a.id}' == articleId, orElse: () {
-      throw 'no article match $lineId.$articleId';
+    final article = calibre.articles.firstWhere(
+        (a) => '${a.calibreId}' == calibreId && '${a.id}' == articleId,
+        orElse: () {
+      throw 'no article match $calibreId.$articleId';
     });
     final ticketsStore = Provider.of<TicketsStore>(context, listen: false);
     final closingStore = Provider.of<ClosingsStore>(context, listen: false);
