@@ -9,34 +9,32 @@ import 'package:rc_router2/rc_router2.dart';
 import 'package:views_weebi/src/articles/article/article_retail/update_a_retail.dart';
 
 class ArticleRetailUpdateRoute extends RcRoute {
-  static String routePath = '/article_retail_update/:lineId/:articleId';
+  static String routePath = '/article_retail_update/:calibreId/:articleId';
 
-  static String generateRoute(String lineId, String articleId) =>
+  static String generateRoute(String calibreId, String articleId) =>
       RcRoute.generateRoute(routePath,
-          pathParams: {'lineId': lineId, 'articleId': articleId});
+          pathParams: {'calibreId': calibreId, 'articleId': articleId});
 
   ArticleRetailUpdateRoute() : super(path: ArticleRetailUpdateRoute.routePath);
 
   @override
   Widget build(BuildContext context) {
     final routeParams = Provider.of<RcRouteParameters>(context);
-    final lineId = routeParams.pathParameters['lineId'] ?? '';
+    final calibreId = routeParams.pathParameters['calibreId'] ?? '';
     final articleId = routeParams.pathParameters['articleId'] ?? '';
     final articlesStore = Provider.of<ArticlesStore>(context, listen: false);
-    final line = articlesStore.lines
-        .firstWhere((p) => p.id.toString() == lineId, orElse: () {
-      throw 'no line match $lineId';
+    final calibre = articlesStore.calibres
+        .firstWhere((p) => p.id.toString() == calibreId, orElse: () {
+      throw 'no calibre match $calibreId';
     });
-    final article = line.articles.firstWhere(
-        (a) => '${a.lineId}' == lineId && '${a.id}' == articleId, orElse: () {
-      throw 'no line match $lineId';
+    final article = calibre.articles.firstWhere(
+        (a) => '${a.calibreId}' == calibreId && '${a.id}' == articleId,
+        orElse: () {
+      throw 'no article retail match $calibreId && $articleId';
     });
 
     return Provider.value(
       value: article,
-      // child: isBasket
-      //     ? ArticleBasketUpdateViewFakeFrame(article as ArticleBasket)
-      //     :
       child: ArticleUpdateView(article as ArticleRetail),
     );
   }
