@@ -3,11 +3,9 @@ library views;
 export 'src/chassis_all_views.dart';
 export 'src/home_view.dart';
 
-export 'src/articles/articles_lines_overview.dart';
+export 'src/articles/overview.dart';
 
-export 'src/articles/line/line_retail/glimpse_line_multiple_articles_r.dart';
-export 'src/articles/line/line_retail/glimpse_line_single_article_r.dart';
-export 'src/articles/line/detail_line.dart';
-export 'src/articles/line/line_retail/tile_title_line_retail.dart';
-
-export 'src/articles/articles_lines_overview.dart';
+export 'src/articles/calibre/retail/calibre_glimpse_multiple_articles.dart';
+export 'src/articles/calibre/retail/calibre_article_single.dart';
+export 'src/articles/calibre/detail_calibre.dart';
+export 'src/articles/calibre/retail/calibre_tile_title.dart';
