@@ -1,15 +1,15 @@
 library article_basket;
 
-export 'src/articles/line/line_basket/create_line_basket.dart';
-export 'src/articles/line/line_basket/glimpse_line_basket.dart';
-export 'src/articles/line/line_basket/tile_title_line_basket.dart';
+export 'src/articles/calibre/basket/calibrate_create_basket.dart';
+export 'src/articles/calibre/basket/calibre_glimpse.dart';
+export 'src/articles/calibre/basket/calibre_tile_title.dart';
 
 export 'src/articles/article/article_basket/update_basket.dart';
 
-export '/src/articles/article/article_basket/type_ahead.dart';
-export '/src/articles/article/article_basket/preview_proxies_selected_widget.dart';
-export '/src/articles/article/article_basket/discount_amount_widget.dart';
-export '/src/articles/article/article_basket/ask_min_qt_basket.dart';
+export 'src/articles/article/article_basket/type_ahead.dart';
+export 'src/articles/article/article_basket/preview_proxies_selected_widget.dart';
+export 'src/articles/article/article_basket/discount_amount_widget.dart';
+export 'src/articles/article/article_basket/ask_min_qt_basket.dart';
 
 export 'src/articles/article/article_basket/detail_section_basket_widget.dart';
 export 'src/articles/article/article_basket/glimpse_a_basket.dart';
